src/backend/requirements.txt
google-api-python-client>=2.150,<3
google-auth>=2.35,<3
google-auth-oauthlib>=1.2,<2
pyairtable>=2.3,<3
google-genai>=1.0,<2
pydantic>=2.10,<3
python-dotenv>=1.0,<2
psycopg2-binary>=2.9,<3
sshtunnel>=0.4,<1
paramiko>=3.5,<4
requests>=2.32,<3
fastapi>=0.115,<1
uvicorn>=0.34,<1
